### 2022-06-20 #4289 @rkurchin
👋 Hi, I noticed this one has been hanging out as `waitlisted` for awhile. I'm by no means an expert in gas chromatography, but I am a Julia user/developer with good expertise in thermodynamics, and decent familiarity with differential equations, so if other editors are tied up, I could take this as I'm about to have a paper accepted and a slot open up (another of my manuscripts is paused, probably for awhile).
