### 2022-07-05 #4289 @rkurchin
👋 @lheagy and @shibabrat would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
