### 2022-07-27 #4565 @bonh
@JanLeppert, I cannot find anything about "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".
