### 2022-07-27 JanLeppert/GasChromatographySimulator.jl#46 @bonh
JOSS: paper 
In reference to https://github.com/openjournals/joss-reviews/issues/4565.

I will post issues and questions about the paper here and hope that we can engage in a discussion about them.
