### 2022-07-27 JanLeppert/GasChromatographySimulator.jl#46 @bonh
Summary: There are a lot of brackets which makes it hard for me to follow. Are they really necessary at that stage?

Statement of need: I think some of the more practical parts in the summary might be better suited in the statement of need. Why gas chromatography? Why simulation of it? And I'm missing "who the target audience is, and its relation to other work".
