### 2022-07-27 JanLeppert/GasChromatographySimulator.jl#46 @bonh
And perhaps you could "mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it."
