### 2022-07-28 #4565 @JanLeppert
> Hi @JanLeppert! So far the packages looks awesome, thanks. I am trying to look into this reference [http://10.1016/j.chroma.2020.460985](http://10.0.3.248/j.chroma.2020.460985) but I cannot download it. Can you provide me a preprint?

Thank you for your comments. Below you can find a copy of the reference you ask for:
[Leppert.2020b-Simulation of spatial thermal gradient gas chromatography-Annotation.pdf](https://github.com/openjournals/joss-reviews/files/9206560/Leppert.2020b-Simulation.of.spatial.thermal.gradient.gas.chromatography-Annotation.pdf)
