### 2022-07-28 JanLeppert/GasChromatographySimulator.jl#46 @JanLeppert
Thank you for your comments. I agree with your notes and I will make some changes to the text in the next days. I also will make some small changes to the documentation due to some updates of the package.
