### 2022-08-02 #4565 @bonh
Thanks, @JanLeppert, for providing the (great!) paper. @rkurchin, I think that this submission together with this [published paper](http://10.0.3.248/j.chroma.2020.460985), see pdf provided by @JanLeppert above, might be a [Co-publication of science, methods, and software](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software). This is somewhat indicated by @JanLeppert in both the JOSS paper and the documentation. I still think that JOSS should consider this submission because in my opinion "the implementation of the software itself reflects a substantial scientific effort" especially due to the extensive and instructive documentation. I think this is an editorial decision, right? So before I continue my review I want to make sure that it is worth the effort.

@JanLeppert, if @rkurchin agrees to move forward with your submission (which I hope!) you should "indicate whether related publications (published, in review, or nearing submission) exist as part of submitting to JOSS", see link above. In my opinion, the paper and this submission are both valuable but perhaps you would like to mention the paper more prominently in your JOSS paper (already in the summary?).

Do produce the results of that paper did you use a predecessor or even an early version of `GasChromatographySimulator.jl`? If you it might make sense to use that paper to provide validating results of your software. Otherwise you might want to add a plot to the JOSS paper indicating what kind of results including their accuracy one can expect from your software.
