### 2022-08-02 #4565 @bonh
@JanLeppert, let's not discuss the points concerning the JOSS paper I raised above here but directly over at https://github.com/JanLeppert/GasChromatographySimulator.jl.
