### 2022-08-02 #4565 @rkurchin
Yes, @bonh, thanks for checking – I agree that this should be fine to publish in JOSS, since the paper is a detailing of the mathematical model but makes very little reference to the software, while this submission is focused on the software and its capabilities. I definitely agree that showing how to reproduce results shown in the theory paper would be excellent content for the project docs, and also that it would be good to reference that paper in your manuscript here, as an example usage (if this package was indeed used for the analysis therein), or at least as related/foundational work!
