### 2022-08-03 JanLeppert/GasChromatographySimulator.jl#50 @ma-sadeghi
JOSS REVIEW: Add community guidelines
I might have missed them, but I couldn't find any community guidelines (how to seek support, where to file issues/bugs, how to contribute to the software). You could probably add a separate section called "Community Guidelines" to the README.
