### 2022-08-03 JanLeppert/GasChromatographySimulator.jl#46 @bonh
I think you should "indicate whether related publications (published, in review, or nearing submission) exist as part of submitting to JOSS", see link above. In my opinion, the paper and this submission are both valuable but perhaps you would like to mention the paper more prominently in your JOSS paper (already in the summary?).
