### 2022-08-03 JanLeppert/GasChromatographySimulator.jl#46 @bonh
To produce the results of the paper http://10.0.3.248/j.chroma.2020.460985 did you use a predecessor or even an early version of GasChromatographySimulator.jl? If you did it might make sense to use that paper to provide validating results of your software. Otherwise you might want to add a plot to the JOSS paper indicating what kind of results including their accuracy one can expect from your software.
