### 2022-08-05 JanLeppert/GasChromatographySimulator.jl#46 @JanLeppert
In the paper was a predecessor used, written in Scilab (similar to Matlab). The fundamental models for fluidics and retention are the same. The Scilab version used a simpler approach to solve the ODE (in principal the Euler method), which toke much more time to run.

I started to implement two examples from this paper (one with conventional GC, one with thermal gradient) in the documentation (#55). I have thought about this earlier. I could than add a plot from these examples (e.g. comparison of measured and simulated chromatograms)
