### 2022-08-05 JanLeppert/GasChromatographySimulator.jl#46 @JanLeppert
In regards to related publications:
- The previous publication [doi.org/10.1016/j.chroma.2021.461943](https://doi.org/10.1016/j.chroma.2021.461943) uses another model (a much more idealized model of chromatography), but in principal uses the same solving of ODEs. For this publication I made the Github project [IBSpublic](https://github.com/JanLeppert/IBSpublic)
- I have some publications in preparation, but they still will take some time
- Based on GasChromatographySimulator.jl I work on two additional packages. One which uses the simulation to calculate the 3 thermodynamic retention parameters of the substances from a handful of temperature programmed measurements using optimization (for this I have a rough prototype, but not yet tested with real data and not optimized). In another project I will use the GasChromatographySimulator.jl to simulate complex GC-systems with multiple GC-columns, e.g. GCxGC, GCxGCxGC, splitting between multiple columns, ... (but this is still in a very early stage)
