### 2022-08-09 JanLeppert/GasChromatographySimulator.jl#50 @JanLeppert
With #55 I added a CONTRIBUTING.md.
