### 2022-08-09 JanLeppert/GasChromatographySimulator.jl#46 @JanLeppert
With #59 I re-worked the draft of the paper.

I changed the order of paragraphs, moved some from section **Summary** to **Statement of need**.

I added a graph, where a measured chromatogram (taken from [10.1016/j.chroma.2020.460985](http://10.0.3.248/j.chroma.2020.460985)) is compared with a simulation. Two comparisons are now part of the documentation in section **Examples**.

I added a paragraph about similar software (lines 55-60) and added another paragraph about related work/projects (line 61-65)
