### 2022-08-09 #4565 @JanLeppert
@bonh , @ma-sadeghi
Thank you for your comments. I made some changes to the package (comments from @ma-sadeghi), added an examples section to the documentation and reworked the draft of the paper.
