### 2022-08-09 JanLeppert/GasChromatographySimulator.jl#60 @ma-sadeghi
JOSS REVIEW: declutter repo root folder
This is just a suggestion, feel free to ignore it if you don't feel like it.

For the sake of decluttering, maybe put all the pictures in the repo root folder in a `bin/` folder (or any other name that suits you).
