### 2022-08-09 JanLeppert/GasChromatographySimulator.jl#61 @ma-sadeghi
JOSS REVIEW: re-generate the figure in the paper with SVG backend
Could you please re-generate the figure in the paper with SVG backend so it looks sharper, currently it's sort of blurry.
