### 2022-08-09 #4565 @ma-sadeghi
@JanLeppert Thank you for your detailed responses. I opened up two more issues ([60](https://github.com/JanLeppert/GasChromatographySimulator.jl/issues/60), [61](https://github.com/JanLeppert/GasChromatographySimulator.jl/issues/61)), they're sort of optional, but I still recommend addressing them.

@rkurchin I'm happy with the responses/actions @JanLeppert has taken (you can find the details in the issue tracker of his package). I've opened up two more issues though. On my end, once those are addressed, I consider this submission to be complete and ready for publication. Please let me know if you need anything else from me.

@JanLeppert Congrats!
