### 2022-08-10 JanLeppert/GasChromatographySimulator.jl#50 @bonh
Did you add this info to the documentation too? I think that would be helpful.
