### 2022-08-10 JanLeppert/GasChromatographySimulator.jl#62 @bonh
JOSS Review: notebooks
Not sure if I am doing something fundamentally wrong but executing the notebook `simulation_conventional_GC.jl` does result in `UndefVarError: DataFrame not defined`.
