### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#62 @JanLeppert
yes, I forgot to update the notebooks, after I changed the reexport (#53). I will change the notebooks and test them.
