### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#60 @JanLeppert
With #63 I moved all files related to the JOSS paper into folder `paper`.
