### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#61 @JanLeppert
With #63 I switched the figure in the JOSS paper from the .png format to the .svg format. It should look better now.
