### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#50 @JanLeppert
With #63 I also added a section `Contribution` to the `README.md`.
