### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#50 @JanLeppert
I also added a section `Contribution` to the first page of the documentation (`dev` version)
