### 2022-08-11 #4565 @JanLeppert
@ma-sadeghi Thank you for your comments. They really helped to improve the package.

I addressed your two additional issues ([60](https://github.com/JanLeppert/GasChromatographySimulator.jl/issues/60), [61](https://github.com/JanLeppert/GasChromatographySimulator.jl/issues/61))
