### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#62 @JanLeppert
Perhaps I could include the notebooks in the `runtests.jl`?
