### 2022-08-11 #4565 @JanLeppert
@bonh Thank you also for your helpful comments. I think, I addressed the two last issues ([62](https://github.com/JanLeppert/GasChromatographySimulator.jl/issues/62), [50](https://github.com/JanLeppert/GasChromatographySimulator.jl/issues/50)). The example notebooks should be functional now and I added the text about contribution also to the documentation.
