### 2022-08-11 #4565 @JanLeppert
@bonh Yes, the name of the 4th author is in greek letters and only the capital greek letters are printed in the pdf. The name is written this way in the exported bibtex-file from the [Zenodo-doi](https://zenodo.org/record/6234110/export/hx#.YvUleexBxhE). I never had this case before.
