### 2022-08-11 JanLeppert/GasChromatographySimulator.jl#62 @bonh
If you want to include the notebooks as tests go ahead. But this does not concern my recommendation for publication in JOSS so I will close this issue.
