### 2022-08-11 #4565 @rkurchin
I haven't seen that issue before, but my first instinct would be to tweak the .bib entry to use the TeX way of specifying Greek characters (i.e. `\Pi \alpha \nu ...`) and see if that fixes it. If that doesn't work, we can tag in someone more familiar with the back end of the typesetting.
