### 2022-08-12 #4565 @JanLeppert
@rkurchin `\alpha` is ignored by the compiler. Using `$\alpha$` prints as `\alpha` in the document. For the greek characters, like , a warning is given by the compiler
`Warning:  Missing character: There is no γ (U+03B3) (U+03B3) in font [lmsans10-regular]:+tlig;!`
The regular font of the paper has no characters for the small greek letters.

Would it be appropriate to change the name to a version using latin characters?
