### 2022-08-12 #4565 @rkurchin
Thanks, reviewers!

@JanLeppert, I've reached out to the JOSS leadership team about this typesetting issue and will let you know how to proceed as soon as I hear back! Thanks for your patience in the meantime.
