### 2022-08-14 #4565 @tarleb
I've added support for Greek letters. It's not perfect, the kerning between uppercase and lowercase chars seems a bit off, but I'm not sure it would be worth it to spend more time on this.
