### 2022-08-14 #4565 @rkurchin
Some editorial comments:
- if you're going to use the acronym GC in the Summary, introduce it after the first use of the phrase (on line 5) and then only use the acronym thereafter. Otherwise, spell it out each time in that section and introduce the acronym in the Statement of Need as you do now
- line 13: "petrochemistry" should be one word
- also line 13: missing (Oxford) comma after "analytics"
- line 15: add comma after "results"
- 20: remove "the" after "models" and hyphenate "one-dimensional"
- 25: change to "temperature `$T(T)$`, programmed inlet pressure `$p_{\text{in}}(t)$`, and programmed flow..."
- 26: change to "...the outlet pressure `$p_{\text{out}}(t)$` can also be programmed."
- 30, 32: should `$t(x)$` actually be $(x,t)$ or something like that? It seems like you're just using it to refer to the coordinate system, but I'm not entirely certain...
- 52: Remove "Fig." since "Figure" already shows up in the citation
- caption of Figure 1: add comma after "details"
- 55: I would replace the phrase "user interface" with "template"
- 57: "allows to simulate" -> "allows simulation of"
- 58-59: "...Available simulations, such as those published in the supplementary material for..."
- 60: hyphenate "free-to-use"
