### 2022-08-15 #4565 @JanLeppert
> * 30, 32: should `$t(x)$` actually be (x,t) or something like that? It seems like you're just using it to refer to the coordinate system, but I'm not entirely certain...

@rkurchin With `$t(x)$` I just wanted to express, that the solution of the first ODE gives the time a substance needs to travel a certain distance. In most cases the movement of an object is described as the distance traveled in a certain time `x(t)`.

I made the other changes you suggested.
